+++ src/write.cpp
@@ -184,7 +184,7 @@ SpatRaster SpatRaster::writeRaster(SpatOptions &opt) {
 				}
 				fnames[i] = out.source[0].filename;
 			}
-			SpatRaster out(fnames, {-1}, {""}, false, {});
+			SpatRaster out(fnames, {-1}, {""}, false, {}, {});
 			return out;
 		}	
 	}