+++ src/write_gdal.cpp
@@ -633,7 +633,7 @@ bool SpatRaster::writeStopGDAL() {
 			GDALClose( (GDALDatasetH) source[0].gdalconnection );
 		} else {
 			GDALClose( (GDALDatasetH) source[0].gdalconnection );
-			GDALDataset *oldDS = openGDAL(copy_filename.c_str(), GDAL_OF_RASTER | GDAL_OF_READONLY);
+			GDALDataset *oldDS = openGDAL(copy_filename.c_str(), GDAL_OF_RASTER | GDAL_OF_READONLY, source[0].open_ops);
 
 			if( oldDS == NULL )  {
 				setError("file copy create failed for "+ copy_driver);